resolved non-determinism issues

+++ src/common.cpp
@@ -109,7 +109,7 @@ boost::thread_specific_ptr<std::string> bundle_label;
 boost::shared_ptr<std::string> bundle_label;
 #endif
 
-int random_seed = -1;
+long random_seed = -1;
 
 extern void print_usage();
